tests/wpt/web-platform-tests/cookies/schemeful-same-site/schemeful-websockets.sub.tentative.html

<!doctype html>
<html>
<head>
  <meta charset=utf-8>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/cookies/resources/testharness-helpers.js"></script>
  <script src="/cookies/resources/cookie-helper.sub.js"></script>
</head>
<body>
<div id=log></div>
<script>
  async_test(function (t) {
    document.cookie = `samesite_strict=1; sameSite=strict; path=/`;
    document.cookie = `samesite_lax=1; sameSite=lax; path=/`;
    t.add_cleanup(function() {
      credFetch(origin + "/cookies/resources/drop.py?name=" + "samesite_strict");
      credFetch(origin + "/cookies/resources/drop.py?name=" + "samesite_lax");
    });


    var ws = new WebSocket("ws://{{host}}:{{ports[ws][0]}}/echo-cookie");
    ws.onclose = t.step_func_done(function () {
      assert_unreached("'close' should not fire before 'open'.");
    });
    ws.onmessage = t.step_func(function (e) {
      ws.onclose = null;
      ws.close();
      // Same-scheme WebSockets should get both cookies
      assert_regexp_match(e.data, /samesite_strict=1/);
      assert_regexp_match(e.data, /samesite_lax=1/);

      var ws2 = new WebSocket("wss://{{host}}:{{ports[wss][0]}}/echo-cookie");
      ws2.onclose = t.step_func_done(function () {
        assert_unreached("'close' should not fire before 'open'.");
      });
      ws2.onmessage = t.step_func_done(function (e2) {
        ws2.onclose = null;
        ws2.close();
        // Cross-scheme WebSockets shouldn't get anything.
        assert_regexp_match(e2.data, /^\(none\)$/, "Cross-scheme");
      });
    });
  }, "Cross-scheme WebSockets are cross-site");
</script>
</body>
</html>
